travellion/requirements.txt
asgiref==3.5.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.0.4
django-allauth==0.54.0
django-cors-headers==3.13.0
django-debug-toolbar==4.1.0
django-filter==22.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
Flask==2.3.3
Flask-Mail==0.9.1
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
Pillow==9.2.0
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==2.0.4
Werkzeug==2.3.7
